+++ src/repobee_plug/cli/io.py
@@ -1,0 +1,10 @@
+"""IO functionality for plugins."""
+
+
+def echo(msg: str) -> None:
+    """Echo a message to the command line.
+
+    Args:
+        msg: The message to echo.
+    """
+    print(msg)
